PRIMO/Func_Data_Transfer/BlobToS3/__init__.py
import logging
import azure.functions as func
from . import Finish_BLOB_S3


def main(req: func.HttpRequest) -> func.HttpResponse:
    logging.info('Python HTTP trigger function processed a request.')

    name = req.params.get('name')
    if not name:
        try:
            data1 = Finish_BLOB_S3.transfer_data()
            print(data1)
            req_body = req.get_json()
        except ValueError:
            pass
        else:
            name = req_body.get('name')

    if name:
        data1 = Finish_BLOB_S3.transfer_data()
        print(data1)
        return func.HttpResponse(f"Hello, {name}. This HTTP triggered function executed successfully.")
    else:
        return func.HttpResponse(
             "This HTTP triggered function executed successfully. Pass a name in the query string or in the request body for a personalized response.",
             status_code=200
        )
